+++ d3fendtools/kuberdf/__init__.py
@@ -1,742 +1,62 @@
-import json
-import logging
-from pathlib import Path
-from time import time
-from typing import Iterable
-from urllib.parse import urlparse
-
-import yaml
-from rdflib import RDF, RDFS, Graph, Literal, Namespace, URIRef, OWL
-
-log = logging.getLogger(__name__)
-K8S = Namespace("urn:k8s:")
-D3F = Namespace("http://d3fend.mitre.org/ontologies/d3fend.owl#")
-NS_DEFAULT = Namespace("https://par-tec.it/example#")
-
-
-def dontyield(*a, **kw):
-    yield from []
-
-
-skip_resource_instances = dontyield
-
-
-SELECTOR_LABELS = ("app.kubernetes.io/name", "app.kubernetes.io/instance", "deployment")
-
-
-class D3fendKube:
-    def __init__(self, text, ontology=None) -> None:
-        self.text = text
-        self.ontology = ontology or Graph()
-        self.annotations = None
-        self.hypergraph = None
-        self.g = Graph()
-        self.g.bind("", NS_DEFAULT)
-        self.g.bind("d3f", D3F)
-        self.g.bind("rdfs", RDFS)
-        self._turtle = None
-        self._out = None
-
-    def parse(self):
-        raise NotImplementedError
-
-    def annotate(self):
-        if self.hypergraph is None:
-            g = Graph()
-            g = self.g | self.ontology
-            annotations = g.query(
-                """
-                PREFIX : <https://par-tec.it/example#>
-
-                CONSTRUCT {
-                ?s a ?parent
-                } WHERE {
-                    ?s a :Node, ?typ .
-                    ?typ rdfs:subClassOf* d3f:DigitalArtifact .
-                    ?typ rdfs:subClassOf ?parent
-                }
-            """,
-                initNs={
-                    "d3f": D3F,
-                },
-            )
-            self.annotations = annotations.graph
-            g = self.g | annotations.graph
-            log.info("Annotated graph has %s triples", len(g) - len(self.g))
-            self.hypergraph = g | self.ontology
-        return self.hypergraph
-
-    def mermaid(self):
-        raise NotImplementedError
-
-
-def strip_oci_image_tag(image: str) -> str:
-    """Strip the tag from an OCI image name"""
-    return image.rsplit(":")[0].split("@sha256", 1)[0]
-
-
-def parse_resources(
-    resources: Iterable[Path], outfile: str, ns_from_file=False, jsonld_output=False
-) -> Graph:
-    g = Graph()
-    g.parse(data=(Path(__file__).parent / "ontology.ttl").read_text(), format="turtle")
-    g.bind("k8s", K8S)
-    for f in resources:
-        ns = f.stem if ns_from_file else None
-        log.info(f"Parsing {f} with namespace {ns}")
-        try:
-            parse_manifest_as_graph(f.read_text(), g=g, manifest_format=f.suffix[1:])
-        except Exception as e:
-            log.exception(f"Error parsing {f}: {e}")
-            raise  # continue
-    dpath = Path(outfile)
-    g.serialize(dpath.with_suffix(".ttl").as_posix(), format="turtle")
-    if jsonld_output:
-        g.serialize(
-            dpath.with_suffix(".jsonld").as_posix(), format="application/ld+json"
-        )
-    return g
-
-
-def parse_manifest(manifest_text: str) -> str:
-    """Parse a manifest text and return a list of triples"""
-    g = parse_manifest_as_graph(manifest_text)
-    return g.serialize(format="turtle")
-
-
-def parse_manifest_as_graph(
-    manifest_text: str, manifest_format="yaml", g=None
-) -> Graph:
-    """Parse a manifest text and return a list of triples.
-    If a graph is provided, the triples will be added to it,
-    otherwise a new graph will be created.
-    """
-
-    def _json_loads_array(json_text):
-        return (json.loads(json_text),)
-
-    ts = time()
-    log.info(f"Reading manifest with format {manifest_format}")
-    manifest_text = manifest_text.strip()
-    if manifest_format == "yaml":
-        parser_f = yaml.safe_load_all
-    elif manifest_format == "json" or manifest_text[0] == "{":
-        parser_f = _json_loads_array
-    else:
-        raise ValueError(f"Unknown manifest format: {manifest_format}")
-    log.warning(f"Read manifest in %d Using parser {parser_f}", (time() - ts))
-    g = g or Graph()
-    g.bind("k8s", K8S)
-    for manifest in parser_f(manifest_text):
-        if not manifest:
-            continue
-        if "kind" not in manifest:
-            log.error(f"Invalid manifest without kind: {manifest}")
-            continue
-        log.error("Parsing manifest %s", manifest.get("kind"))
-        for triple in K8Resource.parse_resource(manifest):
-            g.add(triple)
-    return g
-
-
-def parse_url(url):
-    url = url.strip("jdbc:")
-    if "//" not in url:
-        raise ValueError(f"Invalid URL: {url}")
-    u = urlparse(url)
-    if ":" not in u.netloc:
-        if u.scheme == "mysql":
-            u = u._replace(netloc=f"{u.netloc}:3306")
-        elif u.scheme == "http":
-            u = u._replace(netloc=f"{u.netloc}:80")
-        elif u.scheme == "https":
-            u = u._replace(netloc=f"{u.netloc}:443")
-    return u.netloc
-
-
-class SkipResource:
-    def __init__(self, *a, **k):
-        pass
-
-    triples = dontyield
-
-
-class K8Resource:
-    # Define classmap as a static class constant
-
-    @staticmethod
-    def factory(manifest, ns=None):
-        kind = manifest.get("kind")
-        api_version = manifest.get("apiVersion")
-        log.error(f"Parsing {kind} with apiVersion {api_version} in namespace {ns}")
-        # Use CLASSMAP as the key to fetch the class
-        clz = CLASSMAP.get((api_version, kind)) or K8Resource
-        return clz(manifest, ns=ns)
-
-    @staticmethod
-    def parse_resource(manifest: dict, ns=None):
-        """Parse an OpenShift manifest file
-        and convert it to an RDF resource
-        """
-        resource = K8Resource.factory(manifest, ns=ns)
-        try:
-            if resource.namespace.startswith(("kube-system", "openshift-")):
-                # Ignore kube-system and openshift- namespaces
-                return
-        except AttributeError:
-            pass
-        yield from resource.triples()
-
-    def __init__(self, manifest: dict | None = None, ns: str = None) -> None:
-        if ":" in str(ns):
-            raise ValueError(f"Invalid namespace: {ns}")
-        manifest = manifest or {}
-        self.manifest = manifest
-        self.kind = manifest["kind"]
-        self.metadata = manifest["metadata"]
-        self.name = self.metadata["name"]
-        self.namespace = manifest["metadata"].get("namespace", ns or "default")
-        self.ns = URIRef(f"https://k8s.local/{self.namespace}")
-        self.spec = manifest.get("spec", {})
-        if self.kind == "Namespace":
-            self.uri = self.ns
-        else:
-            self.uri = self.ns + f"/{self.kind}/{self.name}"
-
-        # Set the application.
-        self.app = self.get_app_uri(self.metadata)
-
-    @property
-    def label(self):
-        labelmap = {
-            "BuildConfig": "bc",
-            "Deployment": "dc",
-            "DeploymentConfig": "dc",
-            "ImageStream": "is",
-            "ImageStreamTag": "ist",
-            "Namespace": "ns",
-            "PersistentVolumeClaim": "pvc",
-            "RoleBinding": "rb",
-            "Route": "route",
-            "ServiceAccount": "sa",
-            "StatefulSet": "ss",
-            "Service": "svc",
-        }
-        if self.kind in labelmap:
-            return f"{labelmap[self.kind]}/{self.name}"
-        return f"{self.kind}:{self.name}"
-
-    def triples_kind(self):
-        yield (K8S[self.kind], RDF.type, K8S.Kind)
-
-    def triples_ns(self):
-        yield self.ns, RDF.type, K8S.Namespace
-        yield self.ns, RDFS.label, Literal(self.namespace)
-        yield K8S.cluster, K8S.hasChild, self.ns
-
-    def triples_self(self):
-        yield self.uri, RDF.type, K8S[self.kind]
-        yield self.uri, K8S.hasNamespace, self.ns
-        yield self.ns, K8S.hasChild, self.uri
-        yield self.uri, RDFS.label, Literal(self.label)
-
-        for k, v in self.metadata.get("labels", {}).items():
-            if k not in SELECTOR_LABELS:
-                continue
-            yield self.uri, RDFS.label, Literal(f"{k}: {v}")
-        if self.app:
-            yield self.ns, K8S.hasChild, self.app
-            yield self.app, RDF.type, K8S.Application
-            yield self.app, K8S.hasChild, self.uri
-
-    def get_app_uri(self, metadata):
-        labels = metadata.get("labels", {})
-        app = labels.get("app") or labels.get("application")
-        return URIRef(self.ns + f"/Application/{app}") if app else None
-
-    def triple_spec(self):
-        yield from []
-
-    def triples(self):
-        yield from self.triples_ns()
-        yield from self.triples_self()
-        yield from self.triple_spec()
-
-
-class Route(K8Resource):
-    apiVersion = "route.openshift.io/v1"
-    kind = "Route"
-
-    def triple_spec(self):
-        for k, v in self.spec.items():
-            if k == "host":
-                host_u = URIRef(f"TCP://{v}{self.spec.get('path', '')}:443")
-                yield host_u, RDF.type, K8S.Host
-                yield host_u, D3F.accesses, self.uri
-                yield self.uri, K8S.hasHost, host_u
-                yield host_u, K8S.hasNamespace, self.ns
-                if self.ns:  # FIXME: Add tests
-                    yield self.ns, K8S.hasChild, host_u
-                yield self.ns, K8S.hasChild, host_u
-
-            if k == "to":
-                rel_kind = v["kind"]
-                rel_name = v["name"]
-
-                rel_port = self.spec.get("port", {}).get("targetPort")
-                rel_port = f":{rel_port}" if rel_port else ""
-                yield (
-                    self.uri,
-                    D3F.accesses,
-                    self.ns + f"/{rel_kind}/{rel_name}{rel_port}",
-                )
-                # yield self.ns + f"/{rel_kind}/{rel_name}", RDF.type, NS_K8S[rel_kind]
-                # rel_port = self.spec.get("port", {}).get("targetPort")
-                # dport = URIRef(f"TCP://{rel_name}:{rel_port}")
-                # yield self.uri, NS_K8S.hasTarget, dport
-                # yield dport, RDF.type, NS_K8S.Port
-
-
-class Service(K8Resource):
-    apiVersion = "v1"
-    kind = "Service"
-
-    def triple_spec(self):
-        # Get externalname host or ip
-        if external_name := self.spec.get("externalName"):
-            # ExternalName can cause troubles.
-            # see https://kubernetes.io/docs/concepts/services-networking/service/#externalname
-            host_u = URIRef(f"fixme://{external_name}")
-
-            yield host_u, RDF.type, K8S.Host
-            yield host_u, D3F.accesses, self.uri
-            yield self.uri, K8S.hasHost, host_u
-
-        for port in self.spec.get("ports", []):
-            # Explicit internal TCP connections.
-            port.setdefault("protocol", "TCP")
-
-            host_u = URIRef(f"{port['protocol']}://{self.name}:{port['port']}")
-
-            yield host_u, RDF.type, K8S.Host
-            yield self.uri, K8S.hasHost, host_u
-            if port.get("targetPort"):
-                yield (
-                    self.uri,
-                    K8S.portForward,
-                    Literal("{port}-{protocol}>{targetPort}".format(**port)),
-                )
-            if port_name := port.get("name"):
-                # host_portname_u = URIRef(
-                #     f"{port['protocol']}://{self.name}:{port_name}"
-                # )
-                # yield host_portname_u, RDF.type, NS_K8S.Host
-                # yield self.uri, NS_K8S.hasHost, host_portname_u
-                # yield self.uri, NS_K8S.hasChild, host_portname_u
-                # yield host_portname_u, NS_D3F.accesses, host_u
-                service_port = self.uri + f":{port_name}"
-                yield service_port, RDF.type, K8S.Port
-                yield self.uri, K8S.hasPort, service_port
-                yield self.uri, K8S.hasChild, service_port
-                yield service_port, OWL.sameAs, host_u
-
-            # Service port forwarded to selectors.
-            service_port = self.uri + f":{port['port']}"
-            yield service_port, RDF.type, K8S.Port
-            yield self.uri, K8S.hasPort, service_port
-            yield self.uri, K8S.hasChild, service_port
-            yield host_u, D3F.accesses, service_port
-
-            internal_host_u = URIRef(
-                f"{port['protocol']}://{self.name}.{self.namespace}.svc:{port['port']}"
-            )
-            yield internal_host_u, D3F.accesses, service_port
-            yield self.uri, K8S.hasChild, internal_host_u
-
-            if selector := self.spec.get("selector"):
-                # selector_label = json.dumps(selector, sort_keys=True).replace('"', "")
-                # selector_uuid = hashlib.sha256(selector_label.encode()).hexdigest()
-                # selector_u = URIRef(f"urn:k8s:{self.ns}/Selector/{selector_uuid}")
-                # yield selector_u, RDF.type, NS_K8S.Selector
-                # yield selector_u, RDFS.label, Literal(selector_label)
-                # yield selector_u, NS_K8S.hasNamespace, self.ns
-
-                for k, v in selector.items():
-                    if k not in SELECTOR_LABELS:
-                        continue
-                    # k, v = next(iter(selector.items()))
-                    # Kubernetes by default exposes all ports on a service
-                    endpoint_u = URIRef(
-                        f"{{protocol}}://{self.ns}/{k}={v}:{{targetPort}}".format(
-                            **port
-                        )
-                    )
-                    yield endpoint_u, RDF.type, K8S.Selector
-                    # yield selector_u, NS_K8S.hasChild, endpoint_u
-                    # Service port accesses a selector-based Endpoint.
-                    yield service_port, D3F.accesses, endpoint_u
-                    # yield selector_u, NS_K8S.hasChild, endpoint_u
-            else:
-                # yield an Endpoint with the same name as the service
-                # and on the default namespace.
-                endpoint_u = URIRef(f"urn:k8s:default/Endpoints/{self.name}")
-                yield self.uri, D3F.accesses, endpoint_u
-
-            # Connect host to Endpoint
-            # yield host_u, NS_D3F.accesses, endpoint_u
-            yield self.uri, K8S.hasChild, host_u
-
-
-class DC(K8Resource):
-    """
-    DeploymentConfig
-    $.spec.template.spec.containers[*]
-    """
-
-    @staticmethod
-    def parse_image(image, container_uri=None):
-        image_url_type = len(image.split("/"))
-        if image_url_type == 1:  # only image name
-            image = "https://docker.io/library/" + image
-        elif image_url_type == 2:  # image name and organization
-            image = "https://docker.io/" + image
-        elif image_url_type == 3:  # image name, organization and registry
-            image = "https://" + image
-        else:
-            pass
-
-        image_url = urlparse(image)
-        image_uri = URIRef(
-            image_url._replace(path=strip_oci_image_tag(image_url.path)).geturl()
-        )
-        if image_url.netloc:
-            registry_uri = URIRef(image_url.scheme + "://" + image_url.netloc)
-            yield registry_uri, RDF.type, K8S.Registry
-            yield registry_uri, K8S.hasChild, image_uri
-
-        yield image_uri, RDF.type, K8S.Image
-        if container_uri:
-            yield container_uri, K8S.hasImage, image_uri
-            yield container_uri, D3F.runs, image_uri
-
-    def triple_spec(self):
-        if not (template := self.spec.get("template")):
-            return
-        containers = template.get("spec", {}).get("containers", [])
-        volumes = template.get("spec", {}).get("volumes", [])
-        metadata = template.get("metadata", {})
-        template_labels = metadata.get("labels", {})
-        template_app = self.get_app_uri(metadata) or self.app
-
-        if serviceAccount := template.get("spec", {}).get("serviceAccountName"):
-            service_account = self.ns + f"/ServiceAccount/{serviceAccount}"
-            yield service_account, RDF.type, K8S.ServiceAccount
-            yield self.ns, K8S.hasChild, service_account
-            yield service_account, D3F.executes, self.uri
-
-        for volume in volumes:
-            if "persistentVolumeClaim" in volume:
-                pvc = volume["persistentVolumeClaim"]["claimName"]
-                s_volume = self.ns + f"/PersistentVolumeClaim/{pvc}"
-                yield self.uri, K8S.hasVolume, s_volume
-                yield self.uri, D3F.accesses, s_volume
-                yield s_volume, RDF.type, K8S.PersistentVolumeClaim
-                # XXX: the volume can be mounted in multiple applications.
-                if template_app:
-                    yield template_app, K8S.hasChild, s_volume
-                else:
-                    yield self.ns, K8S.hasChild, s_volume
-            elif "secret" in volume:
-                secret = volume["secret"]["secretName"]
-                s_volume = self.ns + f"/Secret/{secret}"
-                yield self.uri, K8S.hasVolume, s_volume
-                yield self.uri, D3F.reads, s_volume
-                yield s_volume, RDF.type, K8S.Secret
-                yield self.ns, K8S.hasChild, s_volume
-            elif "configMap" in volume:
-                configmap = volume["configMap"]["name"]
-                s_volume = self.ns + f"/ConfigMap/{configmap}"
-                yield self.uri, K8S.hasVolume, s_volume
-                yield self.uri, D3F.reads, s_volume
-                yield s_volume, RDF.type, K8S.ConfigMap
-                yield self.ns, K8S.hasChild, s_volume
-            elif "hostPath" in volume:
-                if self.namespace.startswith(
-                    (
-                        "kube-system",
-                        # "openshift-",
-                        "rook-",
-                    )
-                ):
-                    # Ignore hostPath volumes in kube-system
-                    continue
-                raise NotImplementedError
-            elif "emptyDir" in volume:
-                # Ignore emptyDir volumes
-                continue
-            else:
-                # """
-                # TODO: {'name': 'console-serving-cert',
-                #  'secret': {'defaultMode': 420, 'secretName': 'console-serving-cert'}}
-                #
-                # """
-                raise NotImplementedError(volume)
-
-        for container in containers:
-            s_container = self.uri + f"/Container/{container['name']}"
-            yield self.uri, D3F.runs, s_container
-            yield s_container, RDF.type, K8S.Container
-            yield s_container, RDFS.label, Literal(container["name"])
-            yield self.uri, K8S.hasChild, s_container
-            if self.app:
-                yield self.app, K8S.hasChild, s_container
-
-            if "image" in container:
-                image = DC.parse_image(container["image"], container_uri=s_container)
-                yield from image
-
-            ports = [p.get("containerPort") for p in container.get("ports", [])]
-            ports += [
-                container.get(p, {}).get("httpGet", {}).get("port")
-                for p in ("livenessProbe", "readinessProbe", "startupProbe")
-            ]
-            for port in ports:
-                if not port:
-                    continue
-                protocol = "TCP"
-                for k, v in template_labels.items():
-                    if k not in SELECTOR_LABELS:
-                        continue
-                    port_u = URIRef(f"{protocol}://{self.ns}/{k}={v}:{port}")
-                    yield port_u, RDF.type, K8S.Selector
-                    yield port_u, D3F.accesses, s_container
-                    yield self.uri, K8S.hasChild, port_u
-                    # if template_app:
-                    #     yield template_app, NS_K8S.hasChild, port_u
-
-            for env in container.get("env", []):
-                try:
-                    host_u = parse_url(env["value"])
-                    if "." in host_u:
-                        host_u = URIRef(f"TCP://{host_u}")
-                    else:
-                        # host_u is the name of a kubernetes service
-                        host_u = self.ns + f"/Service/{host_u}"
-
-                    yield host_u, RDF.type, K8S.Host
-                    yield s_container, D3F.accesses, host_u
-                except (
-                    KeyError,
-                    AttributeError,
-                    ValueError,  # parse_url
-                ):
-                    pass
-
-
-class HorizontalPodAutoscaler(K8Resource):
-    """
-    HorizontalPodAutoscaler
-    $.spec.scaleTargetRef
-    """
-
-    apiVersion = "autoscaling/v1"
-    kind = "HorizontalPodAutoscaler"
-
-    def triple_spec(self):
-        if not (scale_target := self.spec.get("scaleTargetRef")):
-            return
-        kind = scale_target["kind"]
-        name = scale_target["name"]
-        if kind in ("DeploymentConfig", "Deployment"):
-            target_u = URIRef(self.ns + f"/{kind}/{name}")
-            yield self.uri, D3F.accesses, target_u
-            yield target_u, K8S.hasChild, self.uri
-            yield target_u, RDF.type, K8S.Job
-            yield self.ns, K8S.hasChild, target_u
-
-
-class ReplicaSet(K8Resource):
-    """
-    ReplicaSet
-    $.spec.template.spec.containers[*]
-    """
-
-    apiVersion = "apps/v1"
-    kind = "ReplicaSet"
-
-    def triples(self):
-        # if it's related to a Deployment, skip it.
-        if self.metadata.get("ownerReferences", [{}])[0].get("kind") == "Deployment":
-            yield from []
-
-
-class CronJob(DC):
-    """
-    CronJob
-    $.spec.jobTemplate.spec.template.spec.containers[*]
-    """
-
-    apiVersion = "batch/v1"
-    kind = "CronJob"
-
-    def __init__(self, *a, **k) -> None:
-        DC.__init__(self, *a, **k)
-        self.spec = self.spec["jobTemplate"]["spec"]
-
-    def triples_spec(self):
-        yield from super().triple_spec()
-        jobs_u = self.ns + "/Jobs"
-        yield jobs_u, K8S.hasChild, self.uri
-
-
-class K8List(K8Resource):
-    apiVersion = "v1"
-    kind = "List"
-
-    def __init__(self, manifest: dict = None, ns: str = None) -> None:
-        """A List is a special resource, don't call super.__init__"""
-        self.kind = manifest["kind"]
-        self.metadata = manifest["metadata"]
-        self.namespace = manifest["metadata"].get("namespace", ns or "default_")
-        self.ns = K8S[self.namespace]
-        self.spec = manifest.get("spec", {})
-        self.items = manifest["items"]
-        # Set the application.
-        self.app = self.get_app_uri(self.metadata)
-
-    def triples(self):
-        for item in self.items:
-            yield from K8Resource.parse_resource(item, ns=self.namespace)
-
-
-class ReplicationController(DC):
-    apiVersion = "v1"
-    kind = "ReplicationController"
-
-    def triples(self):
-        # OCP ReplicationController is created by a DeploymentConfig.
-        if self.metadata.get("annotations", {}).get(
-            "openshift.io/deployment-config.name"
-        ):
-            return
-        # A real ReplicationController.
-        yield from super().triples()
-
-
-class Role(K8Resource):
-    apiVersion = "rbac.authorization.k8s.io/v1"
-    kind = "Role"
-
-
-class RoleBinding(K8Resource):
-    apiVersion = "rbac.authorization.k8s.io/v1"
-    kind = "RoleBinding"
-
-    def triple_spec(self):
-        for subject in self.manifest.get("subjects", []):
-            kind = subject["kind"]
-            name = subject["name"]
-            if namespace := subject.get("namespace"):
-                ns = URIRef(f"https://k8s.local/{namespace}")
-            else:
-                ns = self.ns
-
-            if kind == "ServiceAccount":
-                sa_u = ns + f"/ServiceAccount/{name}"
-                yield sa_u, RDF.type, K8S.ServiceAccount
-                yield ns, RDF.type, K8S.Namespace
-                yield ns, K8S.hasChild, sa_u
-                yield sa_u, K8S.hasNamespace, ns
-                yield self.uri, D3F.authorizes, sa_u
-            elif kind == "User":
-                user_u = URIRef(ns + f"/User/{name}")
-                yield user_u, RDF.type, K8S.User
-                yield ns, K8S.hasChild, user_u
-                yield user_u, K8S.hasNamespace, ns
-                yield self.uri, D3F.authorizes, user_u
-
-            else:
-                raise NotImplementedError(kind)
-
-
-class BuildConfig(K8Resource):
-    apiVersion = "v1"
-    kind = "BuildConfig"
-
-    def triple_spec(self):
-        from_image_url = None
-        to_image_url = None
-
-        to_ = self.spec.get("output", {}).get("to", {})
-        if to_.get("kind") == "ImageStreamTag":
-            to_image_url = URIRef(
-                self.ns + "/ImageStreamTag/" + strip_oci_image_tag(to_["name"])
-            )
-            yield self.uri, K8S.writes, to_image_url
-            yield to_image_url, RDF.type, K8S.ImageStreamTag
-
-        from_ = self.spec.get("strategy", {}).get("sourceStrategy", {}).get("from", {})
-        if from_.get("kind") == "ImageStreamTag":
-            ns = from_.get("namespace", self.ns)
-            from_image_url = URIRef(
-                ns + "/ImageStreamTag/" + strip_oci_image_tag(from_["name"])
-            )
-            yield self.uri, K8S.reads, from_image_url
-            yield from_image_url, RDF.type, K8S.ImageStreamTag
-        if from_image_url and to_image_url:
-            yield to_image_url, RDFS.subClassOf, from_image_url
-
-
-CLASSMAP = {
-    ("kustomize.config.k8s.io/v1beta1", "Kustomization"): SkipResource,
-    ("kustomize.config.k8s.io/v1alpha1", "Component"): SkipResource,
-    ("v1", "Build"): SkipResource,
-    ("v1", "BuildConfig"): BuildConfig,
-    ("v1", "ConsolePlugin"): None,
-    ("batch/v1", "CronJob"): CronJob,
-    ("apps/v1", "Deployment"): DC,
-    ("apps.openshift.io/v1", "DeploymentConfig"): DC,
-    ("v1", "Endpoint"): None,
-    ("v1", "Endpoints"): None,
-    ("autoscaling/v1", "HorizontalPodAutoscaler"): SkipResource,
-    ("image.openshift.io/v1", "ImageStream"): None,
-    ("image.openshift.io/v1", "ImageStreamTag"): None,
-    ("batch/v1", "Job"): SkipResource,
-    ("v1", "List"): K8List,
-    ("v1", "Pod"): SkipResource,
-    ("v1", "Project"): SkipResource,
-    ("v1", "ReplicationController"): ReplicationController,
-    ("route.openshift.io/v1", "Route"): Route,
-    ("v1", "Service"): Service,
-    ("v1", "Secret"): None,
-    ("apps/v1", "StatefulSet"): DC,
-    ("apps/v1", "ReplicaSet"): ReplicaSet,
-    ("autoscaling/v2", "HorizontalPodAutoscaler"): HorizontalPodAutoscaler,
-    ("rbac.authorization.k8s.io/v1", "Role"): Role,
-    ("rbac.authorization.k8s.io/v1", "RoleBinding"): RoleBinding,
-}
-
-# def _register(cls):
-#     """Register a class in the CLASSMAP"""
-#     if not hasattr(cls, "apiVersion") or not hasattr(cls, "kind"):
-#         raise ValueError(f"Class {cls} must have apiVersion and kind attributes")
-#     if not hasattr(cls, "triples"):
-#         raise ValueError(f"Class {cls} must have a triples method")
-#     if not hasattr(cls, "parse_resource"):
-#         raise ValueError(f"Class {cls} must have a parse_resource method")
-#     CLASSMAP[(cls.apiVersion, cls.kind)] = cls
-#     return cls
-#
-# _classes = [
-#     Route,
-#     Service,
-#     DC,
-#     HorizontalPodAutoscaler,
-#     ReplicationController,
-#     CronJob,
-#     K8List,
-#     BuildConfig,
-#     RoleBinding,
-# ]
-# for cls in _classes:
-#     _register(cls)
+from d3fendtools.kuberdf.apicast import (
+    DeveloperAccount,
+    DeveloperUser,
+)
+from d3fendtools.kuberdf.network import (
+    Route,
+    Service,
+)
+from d3fendtools.kuberdf.user import (
+    Role,
+    RoleBinding,
+    BuildConfig,
+)
+from d3fendtools.kuberdf.workload import (
+    Deployment,
+    StatefulSet,
+    CronJob,
+    ReplicaSet,
+    DeploymentConfig,
+    HorizontalPodAutoscaler,
+    ReplicationController,
+)
+from d3fendtools.kuberdf.base import (
+    K8Resource,
+    K8List,
+    K8S,
+    D3F,
+    parse_resources,
+)
+from d3fendtools.kuberdf.externalsecret import (
+    ExternalSecret,
+    PushSecret,
+)
+
+__all__ = [
+    "K8S",
+    "D3F",
+    "K8Resource",
+    "parse_resources",
+    "K8List",
+    # Network
+    "Route",
+    "Service",
+    # User
+    "Role",
+    "RoleBinding",
+    "BuildConfig",
+    # Workload
+    "Deployment",
+    "DeploymentConfig",
+    "HorizontalPodAutoscaler",
+    "ReplicationController",
+    "StatefulSet",
+    "CronJob",
+    "ReplicaSet",
+    # External Secrets
+    "ExternalSecret",
+    "PushSecret",
+    # Apicast
+    "DeveloperAccount",
+    "DeveloperUser",
+]
