+++ d3fendtools/kuberdf/ontology.ttl
@@ -39,6 +39,7 @@ k8s:Service     a rdfs:Class; rdfs:subClassOf k8s:Kind .
 k8s:Node        a rdfs:Class; rdfs:subClassOf k8s:Kind .
 k8s:ServiceAccount a rdfs:Class; rdfs:subClassOf k8s:Kind .
 k8s:RoleBinding a rdfs:Class; rdfs:subClassOf k8s:Kind .
+k8s:ExternalSecret a rdfs:Class; rdfs:subClassOf k8s:Kind .
 
 k8s:Pod rdfs:subClassOf k8s:Kind ;
     rdfs:comment "A group of one or more containers, with shared storage/network, and a specification for how to run the containers" ;
@@ -99,3 +100,4 @@ k8s:Node rdfs:subClassOf d3f:Server .
 k8s:ServiceAccount rdfs:subClassOf d3f:ServiceAccount .
 k8s:RoleBinding rdfs:subClassOf d3f:AccessControlConfiguration .
 k8s:Role rdfs:subClassOf d3f:AccessControlConfiguration .
+k8s:ExternalSecret rdfs:subClassOf d3f:ConfigurationResource, d3f:ScheduledJob .
